requirements.txt
openai==0.28.0
Pillow==10.4.0
PyPDF2==3.0.1
pdfplumber==0.11.3
pytesseract==0.3.10
python-dotenv==1.0.1
langchain==0.2.13